Move TemplateStats and TemplateVariableField
