+++ site/src/pages/TemplatePage/TemplateSummaryPage/TemplateSummaryPageView.tsx
@@ -6,7 +6,7 @@ import {
 import { Loader } from "components/Loader/Loader"
 import { Stack } from "components/Stack/Stack"
 import { TemplateResourcesTable } from "components/TemplateResourcesTable/TemplateResourcesTable"
-import { TemplateStats } from "components/TemplateStats/TemplateStats"
+import { TemplateStats } from "./TemplateStats"
 import { FC, useEffect } from "react"
 import { useLocation, useNavigate } from "react-router-dom"
 import { TemplateVersionWarnings } from "components/TemplateVersionWarnings/TemplateVersionWarnings"
